+++ WorkoutMirrorLive/Features/Live/Feature/LiveFeature.swift
@@ -48,9 +48,11 @@ struct LiveFeature {
                 print("state.destination = nil, from workoutCreatorView")
                 state.destination = nil
                 return .none
-//            case .destination(.presented(.workoutDevicePickerView(..destination(.presented(.workoutTypeListView(.destination(.presented(.workoutSessionView(.summary(.view(.endWorkoutButtonTapped))))))))))):
-//                state.destination = nil
-//                return .none
+            
+            case .destination(.presented(.workoutDevicePickerView(.destination(.presented(.workoutTypeListView(.destination(.presented(.workoutSessionView(.summary(.view(.endWorkoutButtonTapped))))))))))):
+                print("state.destination = nil, from WorkoutSummaryView")
+                state.destination = nil
+                return .none
                 
             case .destination(_):
                 return .none
