+++ WorkoutMirrorLive/Features/Mirroring/Enum/WorkoutSessionStateTest.swift
@@ -1,0 +1,16 @@
+//
+//  WorkoutSessionStateTest.swift
+//  MyFitnessJournal
+//
+//  Created by Sebastian Sciuba on 17/08/2025.
+//
+import Foundation
+
+ //tymczasowo to symulacji
+enum WorkoutSessionStateTest {
+    
+    case running
+//    case ended
+    case paused
+    
+}
