+++ WorkoutMirrorLive/Features/Mirroring/Feature/WorkoutMirroringFeature.swift
@@ -15,6 +15,8 @@ struct WorkoutMirroringFeature {
     
     @Dependency(\.dismiss) var dismiss
     
+    @Dependency(\.workoutSessionClient) var client
+    
     // MARK: - Reducer
     
     var body: some Reducer<State, Action> {
@@ -32,6 +34,13 @@ struct WorkoutMirroringFeature {
                 
             case let .workoutSessionState(value):
                 state.workoutSessionState = value
+                return .none
+                
+            case let .pauseChanged(value):
+                state.isPaused = value
+                if value {
+                    state.pausedElapsedTime = state.elapsedTime
+                }
                 return .none
                 
                 // MARK: - View Action
@@ -66,11 +75,22 @@ struct WorkoutMirroringFeature {
                 state.destination = .heartRateZoneInfo(HeartRateZoneInfoFeature.State())
                 return .none
                 
+            case let .view(.updateElapsedTime(date)):
+                guard !state.isPaused else {
+                    state.elapsedTime = state.pausedElapsedTime
+                    return .none
+                }
+                state.elapsedTime = client.elapsedTimeAt(date)
+                return .none
+                
+            
+//
                 // MARK: - Workout Actions
                 
-            case .view(.pauseWorkoutButtonTaped):
+            case let .view(.pauseWorkoutButtonTaped(value)):
                 return .run { send in
                     await send(.workoutSessionState(.paused))
+                    await send(.pauseChanged(value))
                 }
                 
             case .view(.endWorkoutButtonTapped):
@@ -80,8 +100,13 @@ struct WorkoutMirroringFeature {
 //                }
                 print("endWorkoutButtonTapped")
                 return .none
-            case .view(.resumeWorkoutButtonTapped):
-                return .send(.workoutSessionState(.running))
+                
+            case let .view(.resumeWorkoutButtonTapped(value)):
+                
+                return .run { send in
+                    await send(.workoutSessionState(.running))
+                    await send(.pauseChanged(value))
+                }
                 
             case .view(.xMarkButtonTapped):
                 return .run { send in
@@ -94,6 +119,7 @@ struct WorkoutMirroringFeature {
                 // MARK: - Destination
             case .destination(_):
                 return .none
+      
             }
         }
         .ifLet(\.$destination, action: \.destination)
@@ -101,132 +127,5 @@ struct WorkoutMirroringFeature {
        
 }
 
-/// Implementation of `WorkoutMirroringFeature` action
-extension WorkoutMirroringFeature {
-    
-    @CasePathable
-    enum Action: ViewAction, BindableAction {
-        
-        // MARK: - Binding Action
-        
-        /// Handles changes in bindings for the state.
-        case binding(BindingAction<State>)
-        
-        // MARK: - Action
-        
-        ///
-        case bottomToolBarStateChanged(WorkoutMirroringToolBarState)
-        
-        ///
-        case workoutSessionState(WorkoutSessionStateTest)
-        
-        // MARK: - View Actions
-        
-        case view(View)
-        
-        enum View {
-            
-            /// Action triggered when the view appears on the screen.
-            case viewDidAppear
-            
-            ///
-            case bottomToolBarStateTapped(WorkoutMirroringToolBarState)
-            
-            ///
-            case heartRateZoneButtonTapped
-            
-            ///
-            case hideToolBarButtonTapped
-            
-            // MARK: Camera
-            ///
-            case cameraButtonTapped
-            
-            ///
-            case recordButtonTapped
-            
-            // MARK: Music
-            
-            ////
-            case playPauseMusicButtonTapped
-            
-            ///
-            case forwardMusicButtonTapped
-            
-            ///
-            case backwardMusicButtonTapped
-            
-            // MARK: Workout
-            
-            ///
-            case pauseWorkoutButtonTaped
-            
-            ///
-            case endWorkoutButtonTapped
-            
-            ///
-            case resumeWorkoutButtonTapped
-            
-            ////
-            case xMarkButtonTapped
-        }
-        
-        // MARK: - Navigation and Presentation
-        
-        /// Action to handle navigation destinations within this feature.
-        case destination(PresentationAction<Destination.Action>)
-    }
-}
 
-/// Implementation of `WorkoutMirroringFeature` state
-extension WorkoutMirroringFeature {
-    
-    @ObservableState
-    struct State {
-        
-        // MARK: - Properties
-        
-        var elapsedTime: TimeInterval = 0
-        
-        ///
-        var mirroringToolBarState: WorkoutMirroringToolBarState = .none
-        
-        ///
-        var isActiveCamera: Bool = false
-        
-        ///
-        var recordIsActive: Bool = false
-        
-        ///
-        var isToolbarHidden: Bool = false
-        
-        ///
-        var workoutSessionState: WorkoutSessionStateTest = .running
-        
-        // MARK: - Destination
-        
-        /// destination from WorkoutCreatorFeature
-        @Presents var destination: Destination.State?
-        
-    }
-    
-}
 
-/// Implementation of `WorkoutMirroringFeature` destination
-extension WorkoutMirroringFeature {
-    
-    @Reducer
-    enum Destination {
-        
-        /// Represents the destination for displaying in `HeartRateZoneInfoFeature`.
-        case heartRateZoneInfo(HeartRateZoneInfoFeature)
-        
-    }
-}
-
-// tymczasowo to symulacji
-enum WorkoutSessionStateTest {
-    case running
-//    case ended
-    case paused
-}
